migrated NEventStore for Azure and the core to .net standard 2.0.  In addition, the sql repository was moved out as it has dependencies that do not fit inside of .NET standard.
tests were updated too XUnit 2.  Because of differences in architecture between xunit and xunit 2 there were some substantial changes that had to be made to allow the same BDD test design.  likely more iteration can be done to simplify the testing structure, but not critical for this as all tests still run and perform required functions.